+++ reducers/root-reducer.js
@@ -2,5 +2,6 @@ import React from 'react';
 import { combineReducers } from 'redux'
 import drawings from './drawings';
 import games from './games';
+import frames from './frames';
 
-export default combineReducers({drawings, games});
+export default combineReducers({drawings, games, frames});
